formating

+++ Euler02_linq/Euler02_linq.cs
@@ -8,14 +8,11 @@ namespace Euler02_linqname
     {
         static void Main(string[] args)
         {
-            
-            
-            // Using long just to avoid having to change if we want a higher limit :)
-            IEnumerable<long> Fibonacci()
+            IEnumerable<int> Fibonacci()
             {
-                long current = 0;
-                long next = 1;
-                long temp;
+                int current = 0;
+                int next = 1;
+                int temp;
                 while (true)
                 {
                     yield return current; // returns current no.
@@ -28,9 +25,11 @@ namespace Euler02_linqname
 
             var watch = new System.Diagnostics.Stopwatch();
             watch.Start();
-            long evenSum = Fibonacci().TakeWhile(x => x < 4000000L)  // 'x =>' seems to be iterator
-                          .Where(y => y % 2 == 0)  // second iterator declared here
-                          .Sum();
+            //long evenSum = Fibonacci().TakeWhile(x => x < 4000000L)  // 'x =>' seems to be iterator
+            //              .Where(y => y % 2 == 0)  // second iterator declared here
+            //              .Sum();
+
+            int evenSum = Fibonacci().TakeWhile(x => x < 4000000L).Where(y => y % 2 == 0).Sum();
 
             //Console.WriteLine(evenSum);
             watch.Stop();
